moving fetching logic to the backend
